# app-delivery-backend/src/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API delivery",
    "description": "",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/login/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autentica um usuário existente.",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Credenciais do usuário para login.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "joao123"
                },
                "password": {
                  "type": "string",
                  "example": "senha123"
                }
              },
              "required": [
                "username",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login bem-sucedido.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Login successful"
                },
                "user": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "username": {
                      "type": "string",
                      "example": "joao123"
                    },
                    "is_restaurante": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid credentials"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor durante o login.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal server error during login"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/login/cadastro": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Registra um novo usuário (cliente ou restaurante).",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Dados do usuário para cadastro.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "novo_usuario"
                },
                "email": {
                  "type": "string",
                  "example": "novo@email.com"
                },
                "password": {
                  "type": "string",
                  "example": "senha_segura123"
                },
                "is_restaurante": {
                  "type": "boolean",
                  "example": false
                }
              },
              "required": [
                "username",
                "email",
                "password",
                "is_restaurante"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuário cadastrado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 7
                },
                "username": {
                  "type": "string",
                  "example": "novo_usuario"
                },
                "is_restaurante": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Dados incompletos para o cadastro do usuário.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Todos os campos obrigatorios (username, email, password, is_restaurante) devem ser preenchidos."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor durante o cadastro.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal server error during user registration"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Autentica um usuário existente.",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Credenciais do usuário para login.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "joao123"
                },
                "password": {
                  "type": "string",
                  "example": "senha123"
                }
              },
              "required": [
                "username",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login bem-sucedido.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Login successful"
                },
                "user": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "username": {
                      "type": "string",
                      "example": "joao123"
                    },
                    "is_restaurante": {
                      "type": "boolean",
                      "example": false
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Invalid credentials"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor durante o login.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal server error during login"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/auth/cadastro": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Registra um novo usuário (cliente ou restaurante).",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Dados do usuário para cadastro.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "novo_usuario"
                },
                "email": {
                  "type": "string",
                  "example": "novo@email.com"
                },
                "password": {
                  "type": "string",
                  "example": "senha_segura123"
                },
                "is_restaurante": {
                  "type": "boolean",
                  "example": false
                }
              },
              "required": [
                "username",
                "email",
                "password",
                "is_restaurante"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Usuário cadastrado com sucesso.",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 7
                },
                "username": {
                  "type": "string",
                  "example": "novo_usuario"
                },
                "is_restaurante": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Dados incompletos para o cadastro do usuário.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Todos os campos obrigatorios (username, email, password, is_restaurante) devem ser preenchidos."
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor durante o cadastro.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal server error during user registration"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}